refactor : 이름 수정
BatchConfig -> VoteProcessBatchConfig

VoteProcessingScheduler -> BatchProcessingScheduler
