+++ src/main/java/com/devee/devhive/global/batch/BatchProcessingScheduler.java
@@ -1,7 +1,7 @@
-package com.devee.devhive.domain.project.vote.service;
+package com.devee.devhive.global.batch;
 
 
-import com.devee.devhive.global.config.BatchConfig;
+import com.devee.devhive.global.batch.config.VoteProcessBatchConfig;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
@@ -17,10 +17,10 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class VoteProcessingScheduler {
+public class BatchProcessingScheduler {
 
   private final JobLauncher jobLauncher;
-  private final BatchConfig batchConfig;
+  private final VoteProcessBatchConfig voteProcessBatchConfig;
   private final JobRepository jobRepository;
   private final Step step;
 
@@ -32,7 +32,7 @@ public class VoteProcessingScheduler {
     JobParameters jobParameters = new JobParameters(confMap);
 
     try {
-      jobLauncher.run(batchConfig.voteProcessJob(jobRepository, step), jobParameters);
+      jobLauncher.run(voteProcessBatchConfig.voteProcessJob(jobRepository, step), jobParameters);
     } catch (Exception e) {
       log.info("error: {}", e.getMessage());
     }
